WIP: Move existingCardProcessor to own file
